requirements.txt
Django==3.0
djangorestframework==3.13.1
psycopg2-binary==2.9.1
djangorestframework-simplejwt==5.1.0
